Move latest launch logics from MainContainerActivity to LatestLaunchFragment
