+++ app/src/main/AndroidManifest.xml
@@ -19,9 +19,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".screens.MainContainerActivity">
-
-        </activity>
     </application>
 
 </manifest>
