+++ app/src/main/java/com/nomadworks/spacex/screens/latestlaunch/LatestLaunchFragment.kt
@@ -1,29 +1,41 @@
 package com.nomadworks.spacex.screens.latestlaunch
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
+import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import com.nomadworks.spacex.R
+import com.nomadworks.spacex.databinding.FragmentLatestLaunchBinding
+import dagger.hilt.android.AndroidEntryPoint
 
-class LatestLaunchFragment : Fragment() {
+@AndroidEntryPoint
+class LatestLaunchFragment : Fragment(R.layout.fragment_latest_launch) {
+    private val viewModel: LatestLaunchViewModel by viewModels()
+    private var viewBinding: FragmentLatestLaunchBinding? = null
 
-    companion object {
-        fun newInstance() = LatestLaunchFragment()
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        viewBinding = DataBindingUtil.bind(view)
+        viewBinding?.run {
+            setupViewBinding(this)
+        }
     }
 
-    private lateinit var viewModel: LatestLaunchViewModel
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.latest_launch_fragment, container, false)
+    private fun setupViewBinding(binding: FragmentLatestLaunchBinding) {
+        binding.viewModel = viewModel
+        binding.lifecycleOwner = viewLifecycleOwner
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
+    private fun clearViewBinding(binding: FragmentLatestLaunchBinding) {
+        binding.viewModel = null
     }
 
+    override fun onDestroyView() {
+        super.onDestroyView()
+        viewBinding?.run {
+            clearViewBinding(this)
+        }
+        viewBinding = null
+    }
 }
