+++ app/src/main/res/layout/fragment_latest_launch.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout
     xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     >
     <data>
         <variable
             name="viewModel"
-            type="com.nomadworks.spacex.screens.MainViewModel" />
+            type="com.nomadworks.spacex.screens.latestlaunch.LatestLaunchViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
