+++ build.gradle
@@ -9,9 +9,11 @@ buildscript {
         classpath "com.android.tools.build:gradle:4.0.0"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
+        // Hilt
         classpath "com.google.dagger:hilt-android-gradle-plugin:$hilt_version"
+
+        // Safe Args for Nav component
+        classpath "androidx.navigation:navigation-safe-args-gradle-plugin:$nav_version"
     }
 }
